components/component_test_base: move component_test_base out of test package
This allows external components to use component_test_base and allows
the doc generation to find the package when building from the wheel such
as in our doc_push.sh script.
