+++ examples/apps/dist_cifar/test/component_test.py
@@ -5,7 +5,7 @@
 # LICENSE file in the root directory of this source tree.
 
 import examples.apps.dist_cifar.component as dist_cifar
-from torchx.components.test.component_test_base import ComponentTestCase
+from torchx.components.component_test_base import ComponentTestCase
 
 
 class DistCifar10ComponentTest(ComponentTestCase):
